[refactor/DB-214]: 리뷰 별점 Enum 수정, 산책로 좋아요 수정 (#65)
* refactor(DB-215): infrastructure Rating 수정
* refactor(DB-216): core-domain Rating 수정
* refactor: getWalkwayRating 수정
* refactor(DB-217): api Rating 수정
* feature(DB-218): 별점 검증 어노테이션 추가
* refactor: 산책로 mapper 삭제
* feature(DB-219): 리뷰 작성 가능한 산책로 이용 기록 페이징 추가
* feature(DB-220): 산책로 좋아요 권한 수정
* fix: 산책기록 리뷰 가능 여부 수정
* refactor: 소나 수정
* refactor: 소나 수정
* 서브모듈 업데이트

+++ core-domain/src/main/java/com/dongsan/core/domains/walkway/WalkwayService.java
@@ -61,9 +61,9 @@ public class WalkwayService {
 
     @Transactional
     public void createLikedWalkway(Long memberId, Long walkwayId) {
-        walkwayValidator.validateWalkwayPrivate(walkwayId);
+        Walkway walkway = walkwayReader.getWalkway(walkwayId);
+        walkwayValidator.validateWalkwayAccess(walkway, memberId);
         boolean isLiked = walkwayReader.existsLikedWalkway(memberId, walkwayId);
-
         if (!isLiked) {
             walkwayWriter.saveLikedWalkway(memberId, walkwayId);
         }
@@ -71,9 +71,9 @@ public class WalkwayService {
 
     @Transactional
     public void deleteLikedWalkway(Long memberId, Long walkwayId) {
-        walkwayValidator.validateWalkwayPrivate(walkwayId);
+        Walkway walkway = walkwayReader.getWalkway(walkwayId);
+        walkwayValidator.validateWalkwayAccess(walkway, memberId);
         boolean isLiked = walkwayReader.existsLikedWalkway(memberId, walkwayId);
-
         if (isLiked) {
             walkwayWriter.deleteLikedWalkway(memberId, walkwayId);
         }
@@ -123,9 +123,15 @@ public class WalkwayService {
         return walkwayWriter.saveWalkwayHistory(createWalkwayHistory);
     }
 
-    public List<WalkwayHistory> getCanReviewWalkwayHistory(Long walkwayId, Long memberId) {
+    public List<WalkwayHistory> getCanReviewWalkwayHistory(Long walkwayId, Long memberId, int size, Long lastWalkwayHistoryId) {
         walkwayValidator.validateWalkwayPrivate(walkwayId);
-        return walkwayReader.getCanReviewWalkwayHistory(walkwayId, memberId);
+        LocalDateTime lastCreatedAt = null;
+        if (lastWalkwayHistoryId != null) {
+            WalkwayHistory walkwayHistory = walkwayReader.getWalkwayHistory(lastWalkwayHistoryId);
+            lastCreatedAt = walkwayHistory.createdAt();
+        }
+
+        return walkwayReader.getCanReviewWalkwayHistory(walkwayId, memberId, size, lastCreatedAt);
     }
 
     public List<WalkwayHistory> getUserCanReviewWalkwayHistory(Long memberId, Long lastWalkwayHistoryId, int size) {
@@ -140,6 +146,6 @@ public class WalkwayService {
 
     public boolean isCanReview(Long walkwayHistoryId) {
         WalkwayHistory walkwayHistory = walkwayReader.getWalkwayHistory(walkwayHistoryId);
-        return walkwayHistory.distance() >= walkwayHistory.walkway().courseInfo().distance();
+        return walkwayHistory.distance() >= (walkwayHistory.walkway().courseInfo().distance()) * 2/3;
     }
 }
